examples/time.html

<html>
<h1>Hello, Tizen time API!</h1>

<body>
<pre id="console"></pre>
<script src="js/js-test-pre.js"></script>
<script>
debug('tizen.time.getCurrentDateTime: ' + tizen.time.getCurrentDateTime().toString());
debug('tizen.time.getLocalTimezone: ' + tizen.time.getLocalTimezone());
debug('tizen.time.getAvailableTimezones: ' + tizen.time.getAvailableTimezones());
debug('tizen.time.getDateFormat: ' + tizen.time.getDateFormat());
debug('tizen.time.getTimeFormat: ' + tizen.time.getTimeFormat());

shouldBe("tizen.time.isLeapYear(2013)", "false");
shouldBe("tizen.time.isLeapYear(2014)", "false");
shouldBe("tizen.time.isLeapYear(2015)", "false");
shouldBe("tizen.time.isLeapYear(2016)", "true");
shouldBe("tizen.time.isLeapYear(2017)", "false");
shouldBe("tizen.time.isLeapYear(2018)", "false");
shouldBe("tizen.time.isLeapYear(2019)", "false");
shouldBe("tizen.time.isLeapYear(2020)", "true");
shouldBe("tizen.time.isLeapYear(2021)", "false");
shouldBe("tizen.time.isLeapYear(2022)", "false");
shouldBe("tizen.time.isLeapYear(2023)", "false");

var now = tizen.time.getCurrentDateTime();
debug('Now = ' + now);

var one_day = new tizen.TimeDuration(24, 'HOURS');
debug('One day = ' + one_day);

var tomorrow = now.addDuration(one_day);
debug('Tomorrow = ' + tomorrow);

var two_days = new tizen.TimeDuration(2, 'DAYS');
debug('Two days = ' + two_days);

var minus_one_day = new tizen.TimeDuration(-24, 'HOURS');
var yesterday = now.addDuration(minus_one_day);
debug('Yesterday = ' + yesterday);

var day_after_tomorrow = now.addDuration(two_days);
debug('Day after tomorrow = ' + day_after_tomorrow);

var tomorrow_plus_one_day = tomorrow.addDuration(one_day);
debug('Tomorrow plus one day = ' + tomorrow_plus_one_day);

shouldBe("day_after_tomorrow.equalsTo(tomorrow_plus_one_day)", "true");
shouldBe("day_after_tomorrow.earlierThan(tomorrow_plus_one_day)", "false");
shouldBe("day_after_tomorrow.laterThan(tomorrow_plus_one_day)", "false");

shouldBe("tomorrow.laterThan(now)", "true");
shouldBe("tomorrow.earlierThan(now)", "false");
shouldBe("tomorrow.equalsTo(now)", "false");

shouldBe("yesterday.equalsTo(now)", "false");
shouldBe("now.earlierThan(yesterday)", "false");
shouldBe("now.laterThan(yesterday)", "true");

</script>
<script src="js/js-test-post.js"></script>
</body>
</html>
